fix: move UDP proxy code to utils
